langgraph/mailtoner/pyproject.toml
[project]
name = "mailtoner"
version = "1.0.0"
description = "Change tone of an email"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.3",
    "langchain-google-vertexai>=3.0.1",
    "langgraph>=1.0.2",
    "langgraph-cli[inmem]>=0.4.5",
    "python-dotenv>=1.2.1",
]
